PostNow-REST-API/requirements.txt
# Core Django and DRF
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1

# Authentication and Social Auth
django-allauth==0.61.1
dj-rest-auth==6.0.0
PyJWT==2.10.1

# Database - Vercel-compatible MySQL
PyMySQL==1.1.1
cryptography==45.0.5

# Google OAuth and AI
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.178.0
google-generativeai==0.8.5
google-genai==1.49.0

# AI Services
openai==1.102.0
anthropic==0.64.0

# AWS S3
boto3==1.35.99

# Image Processing
Pillow==11.1.0

# Payment Processing
stripe==12.4.0

# HTTP and API
requests==2.32.4
urllib3==2.5.0

# Utilities
python-dotenv==1.1.0
pydantic==2.11.7

# Vercel deployment
whitenoise==6.6.0

# Mailing API
mailjet-rest==1.5.1

# Celery and Redis for async tasks
celery==5.4.0
redis==5.2.0
django-celery-results==2.5.1
